requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
PyMuPDF>=1.22.0  # For PDF parsing
Pillow>=10.0.0   # For image processing
httpx>=0.24.0    # For HTTP requests
asyncio>=3.4.3   # For async/await support

# Google Generative AI
google-generativeai>=0.3.0

# LightRAG and OpenAI
lightrag>=0.1.0
openai>=1.0.0
numpy>=1.24.0    # For vector operations
scikit-learn>=1.2.0  # For similarity calculations

# Graphiti for knowledge graph
graphiti-core>=0.1.0
neo4j>=5.0.0

# Utilities
tqdm>=4.65.0     # For progress bars
pytest>=7.0.0    # For testing
pytest-asyncio>=0.20.0  # For async tests